+++ config.py
@@ -1,0 +1,9 @@
+from pathlib import Path
+
+
+class Config:
+    MARKDOWN_DIR = Path("World/Ardanos")
+    IMAGE_DIR = Path("World/Images")
+
+
+cfg = Config()