+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
 ### Feat
+* feat: Create sidebar navigation (2025-05-17)
 * feat: Creating DnD wiki project (2025-05-17)